+++ Tests/KeyboardKitTests/Locale/LocaleDictionaryTests.swift
@@ -38,21 +38,24 @@ class LocaleDictionaryTests: QuickSpec {
             )
         }
         
-        it("can be created with any item type") {
-            expect(stringDict).toNot(beNil())
-            expect(intDict).toNot(beNil())
-        }
-        
-        it("can resolve existing values on locale") {
-            let locale = LocaleKey.swedish.locale
-            expect(stringDict.value(for: locale)).to(equal("Swedish"))
-            expect(intDict.value(for: locale)).to(equal(4))
-        }
-        
-        it("returns nil for non-existing locale") {
-            let locale = Locale(identifier: "abc")
-            expect(stringDict.value(for: locale)).to(beNil())
-            expect(intDict.value(for: locale)).to(beNil())
+        describe("locale dictionary") {
+            
+            it("can be created with any item type") {
+                expect(stringDict).toNot(beNil())
+                expect(intDict).toNot(beNil())
+            }
+            
+            it("can resolve existing values on locale") {
+                let locale = LocaleKey.swedish.locale
+                expect(stringDict.value(for: locale)).to(equal("Swedish"))
+                expect(intDict.value(for: locale)).to(equal(4))
+            }
+            
+            it("returns nil for non-existing locale") {
+                let locale = Locale(identifier: "abc")
+                expect(stringDict.value(for: locale)).to(beNil())
+                expect(intDict.value(for: locale)).to(beNil())
+            }
         }
     }
 }